tkinter and button implementation
Added Help button and Place button
Both still print to console

TODO:
Add wipe command that clears all buttons from the field
Replace all other current inputs and error checks with buttons
Draw the line as a label on the top of the program
Replace all print statements with labels that display beneath the buttons?

This is going to require an almost entire rewrite, keeping just the essential parts of the functions. This is going to take a while.