is_empty and useless checks
future proofed is_empty by removing comparison to a gross string and changed it to checking the amount of stones in line.line, what it was doing anyway but now clean.

useless checks to see if the line was empty are removed, because the button to do the action is greyed out if the action is unusable anyway.