Removed code and added highlight
Outdated code and commented block of old code for reference have been removed. References to __repr__ and line.string have also been removed.

Tellstone how has a .highlighted variable. This, when set to True, will replace the regular gray tellstone base with one that has a yellow border for accentuation. This is not currently implimented in any functions yet, just the skeleton. Plans are for this to be added to Peek, Challenge, and Boast