Add highlight to Boast
Also bugfixed highlighting and reveal on multi-peek.

This will most likely be my last commit. I am moving on to learning JavaScript and messing around in my private repositories. That's it for Tellstones 2.0, I'm looking forward to making Tellstones 3.0 as a web app you don't need to download an exe to play. See you soon!